Add protect reinclude update build script

+++ build.sh
@@ -1,12 +1,41 @@
-#!/bin/bash 
-rm bin/* 
-gcc -m32 -c src/int.c    -o bin/int.o
-gcc -m32 -c src/io.c     -o bin/io.o
-gcc -m32 -c src/kernel.c -o bin/kernel.o
-gcc -m32 -c src/stdlib.c -o bin/stdlib.o
-gcc -m32 -c src/string.c -o bin/string.o
+#!/bin/bash
 
-as  --32    src/kernel.s -o bin/conixasm.o
-as  --32    src/irc.s    -o bin/irc.o
+C="gcc -m32 -c"
+LD="ld -m elf_i386 -T"
+SRCF="src"
 
-ld -m elf_i386 -T link.ld -o kernel bin/*
+errCheck(){
+  if (( $? != 0 ))
+  then
+    echo "Error"
+    exit 1
+  fi
+}
+
+cd bin 2>/dev/null
+if  (( $? == 1 ))
+then
+  mkdir -v bin
+else
+  cd ..
+  rm bin/*
+fi
+ls kernel 2>/dev/null >/dev/null
+if (( $? == 0 ))
+then
+  rm kernel
+fi
+
+cd $SRCF
+SRC=( $(ls *.c *.s) )
+cd ..
+
+for ((i=0; i != ${#SRC[@]}; i++))
+do
+  $C $SRCF/${SRC[i]} -o bin/${SRC[i]}.o
+  errCheck
+done
+
+$LD link.ld -o kernel bin/*.o
+errCheck
+echo "Build Complete"
